backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pydantic-settings
fastapi-users[sqlalchemy]
bcrypt
python-multipart
boto3
PyPDF2
python-docx